Many general cleaups and warning fixes. Console code no longer makes assumptions about con_main Screenshots rework, for screenshot_360, but also some other cleanups. Fixed an issue with beginpolygon (finally). Added per-rtlight style strings. Added cvar to control whether ents will be culled by fog. Added define to disable IPLOG, etc. Added r_editlights cvar and related commands, for whenever csaddon isn't available.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5338 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/qclib/qccgui.c
@@ -3012,7 +3012,7 @@ static void EditorReload(editor_t *editor)
 	size_t flensz;
 	char *rawfile;
 	char *file;
-	unsigned int flen;
+	size_t flen;
 	pbool dofree;
 	rawfile = QCC_ReadFile(editor->filename, NULL, NULL, &flensz);
 	flen = flensz;
